+++ Runtime/Utilities/Singleton/SceneSingleton.cs
@@ -1,8 +1,7 @@
 using UnityEngine;
 using NekoLib.Extensions;
-using NekoLib.ColorPalette;
 
-namespace NekoLib.Singleton
+namespace NekoLib.Utilities
 {
     /// <summary>
     /// Singleton that will be destroyed when the scene is unloaded.